slove user_id error and add jwt verification

+++ backend/models/user.go
@@ -3,5 +3,5 @@ package models
 type User struct {
 	UserName string `db:"user_name"`
 	Password string `db:"password"`
-	UserID   int8   `db:"user_id"`
+	UserID   int64  `db:"user_id"`
 }
